Initial Commiit
This assignment focuses on creating a dynamic and interactive webpage using CSS3 transitions and animations, JavaScript functions, and combining CSS animations with JavaScript for enhanced interactivity. The main objective is to build a visually engaging webpage that demonstrates the following:

Create a CSS animation.
Store data in localStorage.
Apply JavaScript to trigger animations.